transcendence/requirements.txt
django==5.0.1
psycopg==3.1.17
argon2-cffi==23.1.0
channels==4.0.0
channels_redis==4.2.0
daphne==4.1.0
pyopenssl==24.0.0
